test: update render result checks
